+++ svg/jvm/src/main/scala/doodle/svg/effect/SvgWriter.scala
@@ -7,13 +7,18 @@ import doodle.effect._
 import doodle.algebra.Picture
 import java.io.File
 import java.nio.file.Files
-import java.util.Base64
+import java.util.{Base64 => JBase64}
 
-object SvgWriter extends Writer[Algebra, Drawing, Frame, Writer.Svg] {
-  def write[A](file: File,
-               description: Frame,
-               picture: Picture[Algebra, Drawing, A]): IO[A] = {
-    Svg.render[Algebra, A](description, algebraInstance, picture)
+object SvgWriter
+    extends Writer[Algebra, Drawing, Frame, Writer.Svg]
+    with Base64[Algebra, Drawing, Frame, Writer.Svg] {
+  def write[A](
+      file: File,
+      description: Frame,
+      picture: Picture[Algebra, Drawing, A]
+  ): IO[A] = {
+    Svg
+      .render[Algebra, A](description, algebraInstance, picture)
       .flatMap {
         case (nodes, a) =>
           IO {
@@ -26,8 +31,13 @@ object SvgWriter extends Writer[Algebra, Drawing, Frame, Writer.Svg] {
   def write[A](file: File, picture: Picture[Algebra, Drawing, A]): IO[A] =
     write(file, Frame("").fitToPicture(), picture)
 
-  override def base64[A](image: Picture[Algebra, Drawing, A]): IO[(A, String)] = for {
-    rendered <- Svg.render[Algebra, A](Frame("").fitToPicture(), algebraInstance, image)
-    (nodes, value) = rendered
-  } yield (value, Base64.getEncoder.encodeToString(nodes.getBytes()))
+  def base64[A](frame: Frame, image: Picture[Algebra, Drawing, A]): IO[(A, String)] =
+    for {
+      rendered <- Svg
+        .render[Algebra, A](frame, algebraInstance, image)
+      (nodes, value) = rendered
+    } yield (value, JBase64.getEncoder.encodeToString(nodes.getBytes()))
+
+  def base64[A](picture: Picture[Algebra, Drawing, A]): IO[(A, String)] =
+    base64(Frame("").fitToPicture(), picture)
 }